refactor(account): 重构账户和交易相关接口及实现
- 将 AbstractAccount 中的 positions 字段访问权限从 private 改为 protected
- 为 AbstractAccount 构造函数添加参数注释文档
- 修改 SimulationAccount 中 enter 和 exit 方法的参数类型从 int 改为 long
- 在 Account 接口中添加 getTotalAmount 方法定义
- 将 enter 方法描述从"买入(做多)"改为"开仓"
- 将 exit 方法描述从"卖出(做空)"改为"平仓"
- 修改 enter 和 exit 方法的 barIdx 参数类型从 int 改为 long
- 在 BarSeries 中为事件回调添加异常处理并调用 System.exit(-1)
- 将 PositionStrategy 中的方法描述从"买入/卖出"改为"开仓/平仓"
- 将 Strategy 中的方法描述从"买入/卖出规则"改为"开仓/平仓规则"
- 将 TradeFeeStrategy 中的方法描述从"买入/卖出手续费"改为"开仓/平仓手续费"
- 新增 TradeListener 接口用于交易监听
- 将 TradeType 枚举描述从"做多/做空"改为"开仓/平仓"
- 从 BaseTest 中移除多余注解并完善测试代码逻辑

+++ clever-task-example/src/main/java/org/clever/quant/account/SimulationAccount.java
@@ -25,13 +25,12 @@ public class SimulationAccount extends AbstractAccount {
     }
 
     @Override
-    public TradeLog enter(int barIdx, double price, long volume, double fee) {
+    public TradeLog enter(long barIdx, double price, long volume, double fee) {
         return null;
     }
 
     @Override
-    public TradeLog exit(int barIdx, double price, long volume, double fee) {
+    public TradeLog exit(long barIdx, double price, long volume, double fee) {
         return null;
     }
 }
-
